Move to uint128 and add vault creation id check

+++ script/utils/ProtocolConfiguration.sol
@@ -49,9 +49,9 @@ abstract contract ProtocolConfiguration is Markets, MarginCollaterals, Sequencer
     uint256 internal constant FINAL_MARGIN_COLLATERAL_ID = 6;
     uint256 internal constant MAX_MARGIN_REQUIREMENTS = 1e18;
     uint256 internal constant MOCK_DATA_STREAMS_EXPIRATION_DELAY = 5 seconds;
-    uint256 internal constant VAULT_ID = 1;
-    uint256 internal constant VAULT_DEPOSIT_CAP = 2e18;
-    uint256 internal constant VAULT_WITHDRAW_DELAY = 1 days;
+    uint128 internal constant VAULT_ID = 1;
+    uint128 internal constant VAULT_DEPOSIT_CAP = 2e18;
+    uint128 internal constant VAULT_WITHDRAW_DELAY = 1 days;
 
 
     /// @notice The maximum delay allowed for the off chain price verification.
